Add is-promise@2.1.0, vue-nav-tabs@^0.5.7, auto-bind^@4.0.0, babel-helper-vue-jsx-merge-props@^2.0.3, style-resources-loader@^1.3.3
Loading specifically 2.1.0 of is-promise because higher version of
is-promise is breaking the build system. Here is the detailed issue:
https://github.com/then/is-promise/issues/13

auto-bind package is used to bind `this` context for class methods when they are
called by event listeners.

style-resources-loader allows to use global/shared scss files

vue-nav-tabs & babel-helper-vue-jsx-merge-props together gives tab ui
feature.